User checkout page

+++ back-end/src/database/models/sale.js
@@ -2,7 +2,7 @@ const Sale = (sequelize, DataTypes) => {
   const Sale = sequelize.define('sale', {
     userId: DataTypes.INTEGER,
     sellerId: DataTypes.INTEGER,
-    totalPrice: DataTypes.DOUBLE,
+    totalPrice: DataTypes.DECIMAL(10,2),
     deliveryAddress: DataTypes.STRING,
     deliveryNumber: DataTypes.STRING,
     saleDate: DataTypes.DATE,
